refactor 001
refactor 001 - create folder structure ( Component, types )
