+++ package.json
@@ -17,6 +17,7 @@
     "next": "13.4.5",
     "react": "18.2.0",
     "react-dom": "18.2.0",
+    "sass": "^1.66.1",
     "typescript": "5.1.3"
   }
 }
