+++ src/app/components/footer/style.scss
@@ -1,0 +1,25 @@
+footer {
+  background-color: #333;
+  color: #fff;
+  text-align: center;
+  position: fixed;
+  left: 0;
+  bottom: 0;
+  width: 100%;
+  padding: 10px 0;
+}
+
+footer a {
+  color: inherit;
+  text-decoration: none;
+}
+
+@media screen and (max-width: 768px) {
+  footer {
+    font-size: 12px;
+  }
+
+  footer p {
+    margin: 10px;
+  }
+}
