+++ src/app/components/header/style.scss
@@ -1,0 +1,5 @@
+header {
+  background-color: #333;
+  color: #fff;
+  padding: 10px;
+}
