+++ src/app/components/home/style.scss
@@ -1,0 +1,101 @@
+.container {
+  text-align: center;
+  margin-top: 50px;
+}
+
+h1 {
+  font-size: 36px;
+  margin-bottom: 20px;
+  color: #333;
+  text-shadow: 1px 1px 2px rgba(0, 0, 0, 0.3);
+}
+
+.subheading {
+  font-size: 1.2rem;
+  font-weight: bold;
+  color: #4a4a4a;
+  margin-bottom: 20px;
+}
+
+.quote-container {
+  background-color: #fff;
+  padding: 40px;
+  border-radius: 10px;
+  box-shadow: 0 2px 6px rgba(0, 0, 0, 0.1);
+}
+
+.quote {
+  font-size: 24px;
+  margin-bottom: 30px;
+  color: #333;
+}
+
+.button-container {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-wrap: wrap;
+  gap: 20px;
+}
+
+.generate-btn,
+.copy-btn,
+.tweet-btn {
+  padding: 10px 20px;
+  font-size: 18px;
+  border-radius: 30px;
+  background-color: #007bff;
+  color: #fff;
+  border: none;
+  cursor: pointer;
+  transition: background-color 0.3s ease;
+}
+
+.generate-btn,
+.copy-btn {
+  background-color: #fe5252;
+}
+
+.tweet-btn:hover {
+  background-color: #0056b3;
+}
+
+.generate-btn:hover,
+.copy-btn:hover {
+  background-color: #af3838;
+}
+
+.tweet-btn {
+  text-decoration: none;
+  display: flex;
+  align-items: center;
+}
+
+.tweet-btn i {
+  margin-right: 5px;
+}
+
+@media screen and (max-width: 768px) {
+  .container {
+    margin-top: 100px;
+  }
+
+  .quote-container {
+    margin: 20px;
+  }
+
+  h1 {
+    font-size: 30px;
+  }
+
+  .quote {
+    font-size: 20px;
+  }
+
+  .generate-btn,
+  .copy-btn,
+  .tweet-btn {
+    font-size: 16px;
+    padding: 8px 16px;
+  }
+}
