+++ src/app/types/home.ts
@@ -1,0 +1,4 @@
+export type quoteType = {
+  text: string;
+  author: string;
+};
